backend/requirements_security.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Werkzeug==2.3.7
PyJWT==2.8.0
pyotp==2.9.0
qrcode[pil]==7.4.2
cryptography==41.0.7
bcrypt==4.0.1
python-dotenv==1.0.0